requirements.txt
Django==1.5.1
distribute==0.6.31
djangorestframework==2.3.6
wsgiref==0.1.2
requests
